OSINT-HUNTER/requirements.txt
requests
bs4
rich
PyPDF2
python-docx
exifread
Pillow
